Update Simple.pde
I think this delay is needed in the other example files, too. Without delay, it can even cause the computer to crash. Since these are examples that are used to understand how the library works, such a thing would be very sad. Also, it really difficult to read the temperature value on the serial monitor. I think this delay makes everyone's life easier.

Thank you!