Use float for calculations between Fahrenheit and Celcius
The result of these calculations is stored in float, which in the case
of 32bit wide floats has 5 digits of accuracy. this is enough for the
purpose. Without this change, the calculations are performed promoted
and performed in double precision, which is very inefficient on some
platforms that do not have support for hardware based double floating point.