Remove 'f'?
HI! I'm working with a temperature Sensor in a lab setting. What does this 'f' do in this code? I have been unable to find its source and what impact f has on the conversion. I started down this road because we need our temperature readings to be more accurate, so I was taking a look at the code. We are using a DS18b20 and I did find the documentation of the sensor in the README file. Our error is closer to 10-15%, and it much higher than the physical sensors error stated in the documentation. Are there ways to calibrate this system with ONEWIRE and DALLASTEMP? I realize this is just a pull request and I really appreciate you taking the time to read.